requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.8
python-dotenv>=1.0
openai>=1.30
tqdm>=4.66
